Workaround Android's lack of javax.lang.model.type.NullType
The NullType class is used as a default value to indicate that
the @Parameters source attribute has not been explicitly set.
As such while it is externally visible it is not something that
clients would be depending upon. Clients are highly unlikely to
either check the source attribute themselves, or specify a value
of Void.class so this change should be safe.

Bug: 30244565
Test: Once it compiles the tests will be run.
Change-Id: If2d1c73e81aabc98f1431206279e82e9e043f4c0