Fix dependency on undefined order of Class.getMethods()
The Class.getMethods() returns an array of methods from a class
and its interfaces and super class but specifies that the order
is undefined. The affected test indexes the array using an index
of 0, presumably because it assumes that the method defined in
the class comes before methods defined in the super class and
interfaces.

This is a bug in the upstream repository and is separated out
from the other patches to make it easier to push it to the
upstream repository.

Bug: 30244565
Change-Id: I9e267b53614a46497e04c73841d086b83b536570
Test: Once it compiles the tests will be run.